requirements.txt
ollama
streamlit
python-dotenv
openai